Directed Unpack Connection

Implementation of ConnectionInterface which represents a directed flow of info between a head module to a tail module. This is an extension of DirectedConnection designed to unpack a specific named value out of a list outputted by the head module as input into the tail module's execution rather than using the direct output of the head module as input into the tail module.

DirectedUnpackConnection

Format

An object of class R6ClassGenerator of length 25.

Class Constructor

new(name, headModule, tailModule, inputArgument)

name: unique character string that identifies this connection.

headModule: obj that implements ModuleInterface or a character string that represents the head module of this connection.

tailModule: obj that implements ModuleInterface or a character string that represents thetail module of this connection.

inputArgument: name of the inputArgument the output of the head module is used for in the tail module's execution. Should be named with what value should be unpacked out of the head module's return list and used for the tail module's execution.

Inheritance From

DirectedConnection

Public

filterOutputValue(outputValue)

Filters the output of the headModule, outputValue, to what is expected by the tailModule. With this implementation of a directed connection, assumes outputValue is a list and the associated name of inputArgument is used to filter down outputValue.

Returns: filtered outputValue

Static Class Methods

DirectedUnpackConnection-cash-initFromSaveData