UpDraft Settings

Stores UpDraft settings. User can directly adjust settings in this object to inject things into UpDraft such as custom logger functions.

UpDraftSettings

Format

An object of class UpDraftSettings (inherits from R6) of length 13.

Adjustable Settings

errorLogger

function to use for UpDraft error statements. Specified function should allow ... string inputs. Caution, specified function should throw a standard R error when called.

infoLogger

function to use for UpDraft info statements. Specified function should allow ... string inputs.

moduleMaxMemorySize

Max memory a module can occupy without causing a runtime error. Value is assumed to be in Bytes.

moduleBorderColor

Graphviz color string of color to render the border of modules in UpDraft workflow visualization.

moduleFillColor

Graphviz color string of color to render the fill of modules in UpDraft workflow visualization.

monitorRate

Numeric value in seconds for the amount of time the monitoring process sleeps between checking if modules have completed.

multipleConnectionsColor

Graphviz color string of color to render multiple connections in UpDraft workflow visualization.

singleConnectionColor

Graphviz color string of color to render single connections in UpDraft workflow visualization.

warnLogger

function to use for UpDraft warning statements. Specified function should allow ... string inputs. Caution, specified function should throw a standard R warning when called.

Public

update(...)

Updates multiple settings at once.

...: named parameters where the name of each parameter is the setting to change and the value of the parameter is the setting's new value.

Returns: NULL

Private

values

Stores values of settings.

setFunction(value, setting)

Updates a setting that stores a function type obj. This is a helper function for this object's active bindings.

value: value to update setting with. If NULL, signifies to return current value of setting.

setting: string of setting to update.

Returns: If value is NULL, returns current setting value. Else, updates setting to value.

setPositiveNumeric(value, setting)

Updates a setting that stores a positive numeric. This is a helper function for this object's active bindings.

value: value to update setting with. If NULL, signifies to return current value of setting.

setting: string of setting to update.

Returns: If value is NULL, returns current setting value. Else, updates setting to value.

setString(value, setting)

Updates a setting that stores a single string. This is a helper function for this object's active bindings.

value: value to update setting with. If NULL, signifies to return current value of setting.

setting: string of setting to update.

Returns: If value is NULL, returns current setting value. Else, updates setting to value.